Added config section to Externals Tools
means config auto-generation stops after the first default generation. (Hidden Flag!)

This addition allows users to select when and what configs should be generated later.
