+++ com.tsk.ide.vscode/Editor/ConfigGeneration.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 63ea0ed645335ce4fa30a10520e0876f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
