+++ com.tsk.ide.vscode/Editor/ConfigGeneration/ConfigGeneration.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d8b7029e9d4497d4b9eded91de2f5acf
+guid: 62a6a6defe5e71a479e0fcaba9e99256
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
