--- com.tsk.ide.vscode/Editor/ProjectGeneration/ProjectGenerationFlag.cs
@@ -1,18 +1,0 @@
-using System;
-
-namespace VSCodeEditor
-{
-    [Flags]
-    public enum ProjectGenerationFlag
-    {
-        None = 0,
-        Embedded = 1,
-        Local = 2,
-        Registry = 4,
-        Git = 8,
-        BuiltIn = 16,
-        Unknown = 32,
-        PlayerAssemblies = 64,
-        LocalTarBall = 128,
-    }
-}
