+++ com.tsk.ide.vscode/Editor/Utils.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8113e8886076ba24c8ff9319fc9a43ce
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
