+++ com.tsk.ide.vscode/Editor/Utils/StringUtils.cs
